Refactor timeline and calendar classes for improved structure and functionality - Updated timeline strings to use singular categories (person, vehicle, animal, package, entity) with labels instead of objects. - Implemented a new API endpoint for creating timeline events.
