+++ custom_components/llmvision/timeline_strings/pl.json
@@ -1,76 +1,64 @@
 {
     "categories": {
-        "people": {
-            "objects": {
-                "osoba": "mdi:walk",
-                "jednostka": "mdi:walk",
-                "postac": "mdi:walk",
-                "ludzie": "mdi:walk",
-                "dziecko": "mdi:walk",
-                "kobieta": "mdi:walk",
-                "mezczyzna": "mdi:walk",
-                "czlowiek": "mdi:walk",
-                "rowerzysta": "mdi:bike"
+        "person": {
+            "labels": {
+                "osoba": "person",
+                "jednostka": "person",
+                "postac": "person",
+                "ludzie": "person",
+                "dziecko": "person",
+                "kobieta": "person",
+                "mezczyzna": "person",
+                "czlowiek": "person"
             }
         },
-        "vehicles": {
-            "objects": {
-                "kurier": "mdi:truck-delivery",
-                "listonosz": "mdi:truck-delivery",
-                "rower": "mdi:bike",
-                "skuter": "mdi:motorbike",
-                "motocykl": "mdi:motorbike",
-                "autobus": "mdi:bus",
-                "samochod": "mdi:car",
-                "dostawczak": "mdi:car",
-                "SUV": "mdi:car",
-                "pojazd": "mdi:car",
-                "ciezarowka": "mdi:truck",
-                "ulica": "mdi:road-variant",
-                "droga": "mdi:road-variant",
-                "podjazd": "mdi:road-variant"
+        "vehicle": {
+            "labels": {
+                "kurier": "truck",
+                "listonosz": "truck",
+                "rower": "bicycle",
+                "skuter": "motorcycle",
+                "motocykl": "motorcycle",
+                "autobus": "bus",
+                "samochod": "car",
+                "dostawczak": "car",
+                "SUV": "car",
+                "pojazd": "car",
+                "ciezarowka": "truck"
             }
         },
-        "animals": {
-            "objects": {
-                "pies": "mdi:dog",
-                "kot": "mdi:cat",
-                "kotek": "mdi:cat",
-                "piesek": "mdi:dog",
-                "ptak": "mdi:duck",
-                "zwierze": "mdi:dog"
+        "animal": {
+            "labels": {
+                "pies": "dog",
+                "kot": "cat",
+                "kotek": "cat",
+                "piesek": "dog",
+                "ptak": "bird"
             }
         },
-        "packages": {
-            "objects": {
-                "pudelko": "mdi:package-variant-closed",
-                "paczka": "mdi:package-variant-closed",
-                "przesylka": "mdi:package-variant-closed",
-                "list": "mdi:email"
+        "package": {
+            "labels": {
+                "pudelko": "package",
+                "paczka": "package",
+                "przesylka": "package"
             }
         },
-        "entities": {
-            "objects": {
-                "kamera": "mdi:cctv",
-                "czujnik": "mdi:access-point",
-                "swiatlo": "mdi:lightbulb",
-                "klucz": "mdi:key",
-                "zamek": "mdi:lock",
-                "alarm": "mdi:alarm-light",
-                "dom": "mdi:home",
-                "weranda": "mdi:door-closed",
-                "drzwi": "mdi:door-closed",
-                "okno": "mdi:window-closed-variant",
-                "brama": "mdi:gate",
-                "bramy": "mdi:gate"
+        "entity": {
+            "labels": {
+                "kamera": "camera",
+                "czujnik": "sensor",
+                "swiatlo": "light",
+                "klucz": "key",
+                "zamek": "lock",
+                "alarm": "alarm"
             }
         },
         "nature": {
-            "objects": {
-                "ogrod": "mdi:flower",
-                "roslina": "mdi:flower",
-                "kwiat": "mdi:flower",
-                "drzewo": "mdi:tree"
+            "labels": {
+                "ogrod": "plant",
+                "roslina": "plant",
+                "kwiat": "plant",
+                "drzewo": "tree"
             }
         }
     },
