+++ custom_components/llmvision/timeline_strings/sk.json
@@ -1,73 +1,64 @@
 {
     "categories": {
-        "people": {
-            "objects": {
-                "osoba": "mdi:walk",
-                "jednotlivec": "mdi:walk",
-                "postava": "mdi:walk",
-                "ľudia": "mdi:walk",
-                "dieťa": "mdi:walk",
-                "žena": "mdi:walk",
-                "muž": "mdi:walk",
-                "človek": "mdi:walk",
-                "cyklista": "mdi:bike"
+        "person": {
+            "labels": {
+                "osoba": "person",
+                "jednotlivec": "person",
+                "postava": "person",
+                "ľudia": "person",
+                "dieťa": "person",
+                "žena": "person",
+                "muž": "person",
+                "človek": "person"
             }
         },
-        "vehicles": {
-            "objects": {
-                "kuriér": "mdi:truck-delivery",
-                "poštár": "mdi:truck-delivery",
-                "bicykel": "mdi:bike",
-                "motocykel": "mdi:motorbike",
-                "motorka": "mdi:motorbike",
-                "autobus": "mdi:bus",
-                "auto": "mdi:car",
-                "dodávka": "mdi:car",
-                "suv": "mdi:car",
-                "vozidlo": "mdi:car",
-                "nákladné auto": "mdi:truck",
-                "ulica": "mdi:road-variant",
-                "chodník": "mdi:road-variant",
-                "príjazdová cesta": "mdi:road-variant"
+        "vehicle": {
+            "labels": {
+                "kuriér": "truck",
+                "poštár": "truck",
+                "bicykel": "bicycle",
+                "motocykel": "motorcycle",
+                "motorka": "motorcycle",
+                "autobus": "bus",
+                "auto": "car",
+                "dodávka": "car",
+                "suv": "car",
+                "vozidlo": "car",
+                "nákladné auto": "truck"
             }
         },
-        "animals": {
-            "objects": {
-                "pes": "mdi:dog",
-                "mačka": "mdi:cat",
-                "vták": "mdi:duck",
-                "domáce zviera": "mdi:dog",
-                "zviera": "mdi:dog"
+        "animal": {
+            "labels": {
+                "pes": "dog",
+                "mačka": "cat",
+                "vták": "bird",
+                "domáce zviera": "dog",
+                "zviera": "dog"
             }
         },
-        "packages": {
-            "objects": {
-                "krabica": "mdi:package-variant-closed",
-                "balík": "mdi:package-variant-closed",
-                "zásielka": "mdi:package-variant-closed",
-                "list": "mdi:email"
+        "package": {
+            "labels": {
+                "krabica": "package",
+                "balík": "package",
+                "zásielka": "package"
             }
         },
-        "entities": {
-            "objects": {
-                "kamera": "mdi:cctv",
-                "senzor": "mdi:access-point",
-                "svetlo": "mdi:lightbulb",
-                "kľúč": "mdi:key",
-                "zámok": "mdi:lock",
-                "alarm": "mdi:alarm-light",
-                "dom": "mdi:home",
-                "veranda": "mdi:door-closed",
-                "dvere": "mdi:door-closed",
-                "okno": "mdi:window-closed-variant"
+        "entity": {
+            "labels": {
+                "kamera": "camera",
+                "senzor": "sensor",
+                "svetlo": "light",
+                "kľúč": "key",
+                "zámok": "lock",
+                "alarm": "alarm"
             }
         },
         "nature": {
-            "objects": {
-                "záhrada": "mdi:flower",
-                "rastlina": "mdi:flower",
-                "kvet": "mdi:flower",
-                "strom": "mdi:tree"
+            "labels": {
+                "záhrada": "plant",
+                "rastlina": "plant",
+                "kvet": "plant",
+                "strom": "tree"
             }
         }
     },
