--- custom_components/llmvision/translations/bg.json
@@ -1,75 +1,0 @@
-{
-    "categories": {
-        "people": {
-            "objects": {
-                "човек": "mdi:walk",
-                "индивид": "mdi:walk",
-                "фигура": "mdi:walk",
-                "хора": "mdi:walk",
-                "дете": "mdi:walk",
-                "жена": "mdi:walk",
-                "мъж": "mdi:walk",
-                "колоездач": "mdi:bike"
-            }
-        },
-        "vehicles": {
-            "objects": {
-                "куриер": "mdi:truck-delivery",
-                "пощальон": "mdi:truck-delivery",
-                "колело": "mdi:bike",
-                "мотор": "mdi:motorbike",
-                "мотоциклет": "mdi:motorbike",
-                "автобус": "mdi:bus",
-                "кола": "mdi:car",
-                "ван": "mdi:car",
-                "джип": "mdi:car",
-                "камион": "mdi:truck",
-                "улица": "mdi:road-variant",
-                "път": "mdi:road-variant",
-                "алея": "mdi:road-variant"
-            }
-        },
-        "animals": {
-            "objects": {
-                "куче": "mdi:dog",
-                "котка": "mdi:cat",
-                "птица": "mdi:duck",
-                "домашен любимец": "mdi:dog",
-                "животно": "mdi:dog"
-            }
-        },
-        "packages": {
-            "objects": {
-                "кутия": "mdi:package-variant-closed",
-                "пакет": "mdi:package-variant-closed",
-                "парцел": "mdi:package-variant-closed",
-                "писмо": "mdi:email"
-            }
-        },
-        "entities": {
-            "objects": {
-                "камера": "mdi:cctv",
-                "сензор": "mdi:access-point",
-                "светлина": "mdi:lightbulb",
-                "крушка": "mdi:lightbulb",
-                "ключ": "mdi:key",
-                "ключалка": "mdi:lock",
-                "аларма": "mdi:alarm-light",
-                "къща": "mdi:home",
-                "дом": "mdi:home",
-                "порта": "mdi:door-closed",
-                "врата": "mdi:door-closed",
-                "прозорец": "mdi:window-closed-variant"
-            }
-        },
-        "nature": {
-            "objects": {
-                "градина": "mdi:flower",
-                "растение": "mdi:flower",
-                "цвете": "mdi:flower",
-                "дърво": "mdi:tree"
-            }
-        }
-    },
-    "regex": "`\\b${key}(а|ът|та|то|те|ите|ове|ове|и|я|ю|е|ен|на|но|ни|ни)?\\b`, 'i'"
-}
